# api/openapi.yaml
openapi: 3.1.0
info:
  title: OnboardMePlease API
  version: 0.4.0
  description: >-
    Capture, resumable synthesis, searchable OKF knowledge and source-grounded questions are implemented.
    Deletion and separate snapshot creation remain planned; those routes are marked below.
servers:
  - url: http://127.0.0.1:8765
    description: Loopback default for the local single-user service.
paths:
...
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/analysis:
    get:
      operationId: analysisStatus
      summary: Latest synthesis job state for this snapshot
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200': { description: Latest job ID and state, or not_started }
        '404': { $ref: '#/components/responses/Problem' }
    post:
      operationId: queueAnalysis
      summary: Analyze up to eight uncached source units and publish assessed overview and OKF
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                cloud_opt_in:
                  type: boolean
                  default: false
                  description: Explicitly saves cloud permission for this repository when installation policy allows it.
      responses:
        '202':
          description: Analysis queued; repeat after completion to resume remaining source units
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobAccepted' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/questions:
    get:
      operationId: getConversation
      summary: Most recent thirty answers in this snapshot, oldest first
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200': { description: Object with answers array }
        '404': { $ref: '#/components/responses/Problem' }
    post:
      operationId: askQuestion
      summary: Retrieve OKF and source, explore bounded relationships, compose and assess a technical answer
      description: Synchronous request with a seven-minute timeout. Mutations require the local session cookie and X-CSRF-Token.
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [question]
              properties:
                question: { type: string, minLength: 1, maxLength: 2000, description: Also limited to 2000 UTF-8 bytes by the server. }
                parent_id: { type: string, description: Previous answer ID from this snapshot, or empty for a new question. }
      responses:
        '200':
          description: Persisted assessed claims, source citations, retrieved OKF concepts, static diagram links and explicit gaps
          content:
            application/json:
              schema:
                $ref: '../contracts/answer.schema.json'
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    RepositoryId:
      name: repositoryId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    SnapshotId:
      name: snapshotId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
  schemas:
    JobAccepted:
      type: object
      additionalProperties: false
      required: [job_id, state]
      properties:
        job_id: { type: string }
        state: { const: queued }
        repository_id: { type: string }
        snapshot_id: { type: string }
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
  responses:
    Problem:
      description: Safe error without raw input, code, path, or credential value
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }